--- a/etl/tempcompare.py
+++ b/etl/tempcompare.py
@@ -20,6 +20,8 @@ def get_list_description_with_max_length(items: List[Any], max_items: int = 20) 
         return ", ".join(str(item) for item in items)
 
 
+
+
 def yield_list_lines(description: str, items: Iterable[Any]) -> Generator[str, None, None]:
     sublines = [item for item in items]
     if len(sublines) > 1: