--- a/lib/catalog/pyproject.toml
+++ b/lib/catalog/pyproject.toml
@@ -31,23 +31,15 @@ rdata = "0.9"
 [tool.poetry.group.dev.dependencies]
 pytest = ">=6.2.5"
 pytest-cov = ">=2.12.1"
-black = "22.3.0"
-flake8 = "6.1.0"
 watchdog = ">=2.1.5"
 argh = ">=0.26.2"
-isort = ">=5.12.0"
 # unpinning those would introduce tons of type errors
 pyright = "1.1.288"
 pandas-stubs = "1.2.0.62"
+ruff = "0.1.6"
 
-
-[tool.isort]
-profile = "black"
-# owid is actually first party library, but we need to stay compatible with ETL
-known_third_party = "owid"
-
-[tool.black]
-line-length = 120
+[tool.ruff]
+extend = "../../pyproject.toml"
 
 [tool.pyright]