--- a/lib/datautils/pyproject.toml
+++ b/lib/datautils/pyproject.toml
@@ -34,8 +34,6 @@ urllib3 = "<2"
 [tool.poetry.group.dev.dependencies]
 pytest = ">=6.2.5"
 pytest-cov = ">=2.12.1"
-black = "22.3.0"
-flake8 = "6.1.0"
 Sphinx = "4.5.0"
 furo = ">=2022.4.7"
 bumpversion = ">=0.6.0"
@@ -46,12 +44,10 @@ ipykernel = ">=6.13.0"
 openpyxl = ">=3.0.10"
 lxml = ">=4.9.1"
 pyright = "1.1.288"
-isort = ">=5.12.0"
+ruff = "0.1.6"
 
-[tool.isort]
-profile = "black"
-
-[tool.black]
+[tool.ruff]
+extend = "../../pyproject.toml"
 
 [tool.pyright]