--- a/lib/repack/Makefile
+++ b/lib/repack/Makefile
@@ -3,7 +3,7 @@
 #
 
 
-include default.mk
+include ../../default.mk
 
 SRC = owid tests