--- a/lib/repack/pyproject.toml
+++ b/lib/repack/pyproject.toml
@@ -17,17 +17,10 @@ pandas = ">=1.5.2"
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.2.0"
 pyright = ">=1.1.285"
-black = "22.3.0"
-flake8 = "6.1.0"
-isort = ">=5.11.4"
+ruff = "0.1.6"
 
-[tool.isort]
-profile = "black"
-# owid is actually first party library, but we need to stay compatible with ETL
-known_third_party = "owid"
-
-[tool.black]
-line-length = 120
+[tool.ruff]
+extend = "../../pyproject.toml"
 
 [tool.pyright]