--- a/lib/repack/tests/test_repack.py
+++ b/lib/repack/tests/test_repack.py
@@ -3,8 +3,9 @@ from typing import Any
 import numpy as np
 import pandas as pd
 import pytest
-from owid import repack
 from pandas.testing import assert_frame_equal
+
+from owid import repack
 
 
 def test_repack_non_object_columns():