--- a/pyproject.toml
+++ b/pyproject.toml
@@ -89,7 +89,6 @@ selenium = "^4.15.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.1.2"
-flake8 = "6.1.0"
 watchdog = ">=2.1.5"
 PyYAML = ">=5.4.1"
 argh = ">=0.26.2"
@@ -98,11 +97,9 @@ ipython = ">=7.27.0"
 matplotlib = ">=3.4.3"
 plotly = ">=5.7.0"
 cookiecutter = ">=2.1.1"
-isort = ">=5.12.0"
 types-PyYAML = ">=6.0.12"
 hydra-core = ">=1.3.0"
 pywebio = ">=1.8.2"
-black = {version = "22.3.0", extras = ["jupyter"]}
 boto3-stubs = { extras = ["s3"], version = ">=1.21.16" }
 gspread = "^5.10.0"
 # unpinning those would introduce tons of type errors
@@ -112,15 +109,21 @@ jsonref = "^1.1.0"
 mkdocs-jupyter = "^0.24.2"
 mkdocs-exclude = "^1.0.2"
 mkdocs-gen-files = "^0.5.0"
+ruff = "0.1.6"
 
-
-[tool.black]
-exclude = ".ipynb_checkpoints|apps/wizard/templating/cookiecutter|.*playground.*.ipynb"
+[tool.ruff]
+extend-select = [
+    # isort
+    "I"
+]
+ignore = ["E501"]
 line-length = 120
+target-version = "py310"
+extend-exclude = [
+    ".ipynb_checkpoints",
+    "*cookiecutter",
+]
 
-[tool.isort]
-profile = "black"
-extend_skip = [".ipynb_checkpoints"]
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]