The ruff linter is a much faster linter written in rust.