@Marigold 2023-08-10
How far did you get @larsyencken? I'm having minor issues locally with isort and flake8 and this might perhaps solve them both.