@Marigold 2023-11-20
Speedup:

* isort -> ruff: 7.8s -> 1.6s
* black -> ruff: 6s -> 1.8s