@larsyencken 2023-11-20
Nice! What's the best way to get it working in VSCode? It's a pet hate of mine when CI and editor lint differently.