@Marigold 2023-11-20
Sorry! It was a WIP comment and I accidentally submitted it. I'm preparing more thorough overview (plus of course how to set it up in VSCode). It looks pretty promising so far!