@Marigold 2023-11-20
@pabloarosado here's a [summary](https://github.com/owid/etl/pull/1017#issuecomment-1818665033) of the proposed change. What do you think, is it worth it or would you rather stick to our existing tools?