@larsyencken 2023-11-20
> Looks good, and the speed up is going to be very much appreciated! I haven't tried it, have you noticed any major change in how files are formatted? If not, I'm happy to switch to ruff.

Ruff tries to format identically to Black on pre-blackened code, though on code that hasn't been auto-formatted it might make a different choice. https://docs.astral.sh/ruff/formatter/#black-compatibility